requirements.txt
pyyaml>=6.0
openai>=1.0.0
tqdm>=4.65.0
click>=8.0.0 